+++ test/FastRouteRouterTest.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-fastroute for the canonical source repository
- * @copyright Copyright (c) 2015-2016 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2015-2017 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-fastroute/blob/master/LICENSE.md New BSD License
  */
 
@@ -10,6 +10,7 @@ namespace ZendTest\Expressive\Router;
 use FastRoute\Dispatcher\GroupCountBased as Dispatcher;
 use FastRoute\RouteCollector;
 use PHPUnit\Framework\TestCase;
+use Prophecy\Prophecy\ProphecyInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\UriInterface;
 use Zend\Expressive\Router\Exception\InvalidArgumentException;
@@ -19,11 +20,26 @@ use Zend\Expressive\Router\RouteResult;
 
 class FastRouteRouterTest extends TestCase
 {
+    /**
+     * @var RouteCollector|ProphecyInterface
+     */
+    private $fastRouter;
+
+    /**
+     * @var Dispatcher|ProphecyInterface
+     */
+    private $dispatcher;
+
+    /**
+     * @var callable
+     */
+    private $dispatchCallback;
+
     public function setUp()
     {
         $this->fastRouter = $this->prophesize(RouteCollector::class);
         $this->dispatcher = $this->prophesize(Dispatcher::class);
-        $this->dispatchCallback = function ($data) {
+        $this->dispatchCallback = function () {
             return $this->dispatcher->reveal();
         };
     }
@@ -161,6 +177,8 @@ class FastRouteRouterTest extends TestCase
 
     /**
      * @depends testMatchingRouteShouldReturnSuccessfulRouteResult
+     *
+     * @param array $data
      */
     public function testMatchedRouteResultContainsRoute(array $data)
     {
@@ -179,6 +197,8 @@ class FastRouteRouterTest extends TestCase
 
     /**
      * @dataProvider idemPotentMethods
+     *
+     * @param string $method
      */
     public function testRouteNotSpecifyingOptionsImpliesOptionsIsSupportedAndMatchesWhenGetOrHeadIsAllowed(
         $method
@@ -364,6 +384,10 @@ class FastRouteRouterTest extends TestCase
      * @group zendframework/zend-expressive#53
      * @group 8
      * @dataProvider generatedUriProvider
+     *
+     * @param array $routes
+     * @param string $expected
+     * @param array $generateArgs
      */
     public function testCanGenerateUriFromRoutes(array $routes, $expected, array $generateArgs)
     {
@@ -424,18 +448,6 @@ class FastRouteRouterTest extends TestCase
         $this->assertEquals('foo-route', $result->getMatchedRouteName());
     }
 
-    public function testGenerateUriRaisesExceptionForIncompleteUriSubstitutions()
-    {
-        $router = new FastRouteRouter();
-        $route = new Route('/foo[/{param}[/optional-{extra}]]', 'foo', ['GET'], 'foo');
-        $router->addRoute($route);
-
-        $this->expectException(InvalidArgumentException::class);
-        $this->expectExceptionMessage('unsubstituted parameters');
-
-        $router->generateUri('foo', ['extra' => 'segment']);
-    }
-
     public function uriGeneratorDataProvider()
     {
         return [
@@ -457,8 +469,11 @@ class FastRouteRouterTest extends TestCase
 
     /**
      * @dataProvider uriGeneratorDataProvider
+     *
+     * @param string $expectedUri
+     * @param array $params
      */
-    public function testUriGenerationSubstitutionsWithDefaultOptions($expectedUri, $params)
+    public function testUriGenerationSubstitutionsWithDefaultOptions($expectedUri, array $params)
     {
         $router = new FastRouteRouter();
 
@@ -477,8 +492,11 @@ class FastRouteRouterTest extends TestCase
 
     /**
      * @dataProvider uriGeneratorDataProvider
+     *
+     * @param string $expectedUri
+     * @param array $params
      */
-    public function testUriGenerationSubstitutionsWithDefaultsAndOptionalParameters($expectedUri, $params)
+    public function testUriGenerationSubstitutionsWithDefaultsAndOptionalParameters($expectedUri, array $params)
     {
         $router = new FastRouteRouter();
 
@@ -517,8 +535,11 @@ class FastRouteRouterTest extends TestCase
 
     /**
      * @dataProvider uriGeneratorWithPartialDefaultsDataProvider
+     *
+     * @param string $expectedUri
+     * @param array $params
      */
-    public function testUriGenerationSubstitutionsWithPartialDefaultsAndOptionalParameters($expectedUri, $params)
+    public function testUriGenerationSubstitutionsWithPartialDefaultsAndOptionalParameters($expectedUri, array $params)
     {
         $router = new FastRouteRouter();
 
@@ -597,5 +618,20 @@ class FastRouteRouterTest extends TestCase
         $this->assertEquals('fooHandler', $result->getMatchedMiddleware());
 
         unlink($cache_file);
+    }
+
+    /**
+     * Test for issue #30
+     */
+    public function testGenerateUriRaisesExceptionForMissingMandatoryParameters()
+    {
+        $router = new FastRouteRouter();
+        $route = new Route('/foo/{id}', 'foo', ['GET'], 'foo');
+        $router->addRoute($route);
+
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('expects at least parameter values for');
+
+        $router->generateUri('foo');
     }
 }